Refactor ics & vobject helpers into gcalcli.ics
