+++ stubs/vobject/__init__.pyi
@@ -1,0 +1,4 @@
+from .base import Component, readComponents as readComponents
+
+def iCalendar() -> Component: ...
+def vCard() -> Component: ...
