+++ tox.ini
@@ -24,6 +24,7 @@ description = run type checks
 skip_install = true
 deps =
     mypy >= 1.0
+    google-api-python-client-stubs
     types-python-dateutil
     types-requests
     types-vobject
